Initial transform-gizmo structure changes
