+++ .gitignore
@@ -1,4 +1,5 @@
 /target
 Cargo.lock
 /.idea
+/.vscode
 .DS_Store
