+++ crates/transform-gizmo-demo/Trunk.toml
@@ -1,0 +1,4 @@
+[build]
+public_url = "./"
+filehash = false
+dist = "../docs"
