--- demo/.gitignore
@@ -1,4 +1,0 @@
-/target
-Cargo.lock
-/.idea
-.DS_Store
