--- demo/index.html
@@ -1,21 +1,0 @@
-<html>
-<head>
-    <title>egui-gizmo demo</title>
-    <style>
-        html,
-        body,
-        canvas {
-            margin: 0;
-            padding: 0;
-            width: 100%;
-            height: 100%;
-            overflow: hidden;
-            position: absolute;
-            background: black;
-            z-index: 0;
-        }
-    </style>
-</head>
-<body oncontextmenu="return false;"> <!-- Disable right-click menu, as it right-click is used for rotation -->
-</body>
-</html>
