--- docs/egui-gizmo-demo.js
@@ -1,1 +1,0 @@
-let a5=4,W=0,T=null,Q=`undefined`,a0=`boolean`,_=`function`,a1=`string`,a9=42857,Y=1,a2=`Object`,U=`utf-8`,$=`number`,S=Array,V=Error,a4=Float32Array,Z=Int32Array,a3=JSON.stringify,a7=Object,a8=Reflect,a6=Uint32Array,X=Uint8Array,R=undefined;var F=((a,b)=>{a=a>>>W;return E().subarray(a/a5,a/a5+ b)});var v=(a=>{const b=typeof a;if(b==$||b==a0||a==T){return `${a}`};if(b==a1){return `"${a}"`};if(b==`symbol`){const b=a.description;if(b==T){return `Symbol`}else{return `Symbol(${b})`}};if(b==_){const b=a.name;if(typeof b==a1&&b.length>W){return `Function(${b})`}else{return `Function`}};if(S.isArray(a)){const b=a.length;let c=`[`;if(b>W){c+=v(a[W])};for(let d=Y;d<b;d++){c+=`, `+ v(a[d])};c+=`]`;return c};const c=/\[object ([^\]]+)\]/.exec(toString.call(a));let d;if(c.length>Y){d=c[Y]}else{return toString.call(a)};if(d==a2){try{return `Object(`+ a3(a)+ `)`}catch(a){return a2}};if(a instanceof V){return `${a.name}: ${a.message}\n${a.stack}`};return d});var E=(()=>{if(D===T||D.byteLength===W){D=new a4(b.memory.buffer)};return D});var M=((a,b)=>{});var y=((a,c,d)=>{b._dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h067ac15b54489931(a,c,l(d))});function C(a,c){try{return a.apply(this,c)}catch(a){b.__wbindgen_exn_store(l(a))}}var m=(a=>a===R||a===T);var d=(a=>c[a]);var o=(()=>{if(n===T||n.byteLength===W){n=new Float64Array(b.memory.buffer)};return n});var L=(()=>{const c={};c.wbg={};c.wbg.__wbg_instanceof_Response_0d25bb8436a9cefe=(a=>{let b;try{b=d(a) instanceof Response}catch(a){b=!1}const c=b;return c});c.wbg.__wbindgen_object_drop_ref=(a=>{g(a)});c.wbg.__wbg_stringify_daa6661e90c04140=function(){return C((a=>{const b=a3(d(a));return l(b)}),arguments)};c.wbg.__wbg_fetch_0117c27c9b3739e0=((a,b,c)=>{const e=d(a).fetch(k(b,c));return l(e)});c.wbg.__wbg_status_351700a30c61ba61=(a=>{const b=d(a).status;return b});c.wbg.__wbg_arrayBuffer_ec4617b29bb0f61c=function(){return C((a=>{const b=d(a).arrayBuffer();return l(b)}),arguments)};c.wbg.__wbg_new_ace717933ad7117f=(a=>{const b=new X(d(a));return l(b)});c.wbg.__wbg_mark_40e050a77cc39fea=((a,b)=>{performance.mark(k(a,b))});c.wbg.__wbg_log_c9486ca5d8e2cbe8=((a,c)=>{let d;let e;try{d=a;e=c;console.log(k(a,c))}finally{b.__wbindgen_free(d,e,Y)}});c.wbg.__wbg_log_aba5996d9bde071f=((a,c,d,e,f,g,h,i)=>{let j;let l;try{j=a;l=c;console.log(k(a,c),k(d,e),k(f,g),k(h,i))}finally{b.__wbindgen_free(j,l,Y)}});c.wbg.__wbg_width_9d9d26b087c6ad54=(a=>{const b=d(a).width;return b});c.wbg.__wbg_height_770da314320603d8=(a=>{const b=d(a).height;return b});c.wbg.__wbg_matchMedia_fed5c8e73cf148cf=function(){return C(((a,b,c)=>{const e=d(a).matchMedia(k(b,c));return m(e)?W:l(e)}),arguments)};c.wbg.__wbg_matches_68b7ad47c1091323=(a=>{const b=d(a).matches;return b});c.wbg.__wbg_removeEventListener_c62be4990a8c9c02=function(){return C(((a,b,c,e,f)=>{d(a).removeEventListener(k(b,c),d(e),d(f))}),arguments)};c.wbg.__wbindgen_string_new=((a,b)=>{const c=k(a,b);return l(c)});c.wbg.__wbg_error_0f3a2d4325dee96a=((a,b)=>{console.error(d(a),d(b))});c.wbg.__wbindgen_object_clone_ref=(a=>{const b=d(a);return l(b)});c.wbg.__wbg_addEventListener_1b158e9e95e0ab00=function(){return C(((a,b,c,e,f)=>{d(a).addEventListener(k(b,c),d(e),d(f))}),arguments)};c.wbg.__wbg_addEventListener_2f891d22985fd3c8=function(){return C(((a,b,c,e)=>{d(a).addEventListener(k(b,c),d(e))}),arguments)};c.wbg.__wbg_new_87d841e70661f6e9=(()=>{const a=new a7();return l(a)});c.wbg.__wbg_document_5257b70811e953c0=(a=>{const b=d(a).document;return m(b)?W:l(b)});c.wbg.__wbg_querySelector_d86f889797c65e88=function(){return C(((a,b,c)=>{const e=d(a).querySelector(k(b,c));return m(e)?W:l(e)}),arguments)};c.wbg.__wbg_parentElement_86a7612dde875ba9=(a=>{const b=d(a).parentElement;return m(b)?W:l(b)});c.wbg.__wbg_getBoundingClientRect_f3f6eb39f24c1bb0=(a=>{const b=d(a).getBoundingClientRect();return l(b)});c.wbg.__wbg_width_164c11c1f72aa632=(a=>{const b=d(a).width;return b});c.wbg.__wbg_height_ac60120008caa50b=(a=>{const b=d(a).height;return b});c.wbg.__wbindgen_cb_drop=(a=>{const b=g(a).original;if(b.cnt--==Y){b.a=W;return !0};const c=!1;return c});c.wbg.__wbg_instanceof_HtmlCanvasElement_a6076360513b6876=(a=>{let b;try{b=d(a) instanceof HTMLCanvasElement}catch(a){b=!1}const c=b;return c});c.wbg.__wbg_stopPropagation_786ab850031995e5=(a=>{d(a).stopPropagation()});c.wbg.__wbg_cancelBubble_191799b8e0ab3254=(a=>{const b=d(a).cancelBubble;return b});c.wbg.__wbg_preventDefault_d2c7416966cb0632=(a=>{d(a).preventDefault()});c.wbg.__wbg_matches_2a7b0d97653c323f=(a=>{const b=d(a).matches;return b});c.wbg.__wbg_target_791826e938c3e308=(a=>{const b=d(a).target;return m(b)?W:l(b)});c.wbg.__wbg_is_a5728dbfb61c82cd=((a,b)=>{const c=a7.is(d(a),d(b));return c});c.wbg.__wbg_offsetX_72778aba0c4b5674=(a=>{const b=d(a).offsetX;return b});c.wbg.__wbg_offsetY_a8d9605d7c28379d=(a=>{const b=d(a).offsetY;return b});c.wbg.__wbg_requestFullscreen_86cc33835b4db656=function(){return C((a=>{d(a).requestFullscreen()}),arguments)};c.wbg.__wbg_movementX_7ed3fefa16dfa971=(a=>{const b=d(a).movementX;return b});c.wbg.__wbg_movementY_a0be141073121d2c=(a=>{const b=d(a).movementY;return b});c.wbg.__wbg_buttons_d516d4a6ffb63df2=(a=>{const b=d(a).buttons;return b});c.wbg.__wbg_pointerType_6421ba54876364b9=((a,c)=>{const e=d(c).pointerType;const f=u(e,b.__wbindgen_malloc,b.__wbindgen_realloc);const g=r;q()[a/a5+ Y]=g;q()[a/a5+ W]=f});c.wbg.__wbg_pointerId_288a7753a42433eb=(a=>{const b=d(a).pointerId;return b});c.wbg.__wbg_clientX_4d37584813a1790a=(a=>{const b=d(a).clientX;return b});c.wbg.__wbg_clientY_ea543e0b8dc1490d=(a=>{const b=d(a).clientY;return b});c.wbg.__wbg_pressure_ef807a4027b5b179=(a=>{const b=d(a).pressure;return b});c.wbg.__wbg_key_9a2550983fbad1d0=((a,c)=>{const e=d(c).key;const f=u(e,b.__wbindgen_malloc,b.__wbindgen_realloc);const g=r;q()[a/a5+ Y]=g;q()[a/a5+ W]=f});c.wbg.__wbg_ctrlKey_e7fc1575581bc431=(a=>{const b=d(a).ctrlKey;return b});c.wbg.__wbg_altKey_c3c61dc3af936846=(a=>{const b=d(a).altKey;return b});c.wbg.__wbg_getModifierState_3adb04889097105e=((a,b,c)=>{const e=d(a).getModifierState(k(b,c));return e});c.wbg.__wbg_keyCode_6acbcd0e4e062504=(a=>{const b=d(a).keyCode;return b});c.wbg.__wbg_charCode_7bf774ca7e2117ac=(a=>{const b=d(a).charCode;return b});c.wbg.__wbg_setPointerCapture_02adb3c41a2a5367=function(){return C(((a,b)=>{d(a).setPointerCapture(b)}),arguments)};c.wbg.__wbg_requestAnimationFrame_1820a8e6b645ec5a=function(){return C(((a,b)=>{const c=d(a).requestAnimationFrame(d(b));return c}),arguments)};c.wbg.__wbg_setTimeout_bd20251bb242e262=function(){return C(((a,b,c)=>{const e=d(a).setTimeout(d(b),c);return e}),arguments)};c.wbg.__wbg_body_3eb73da919b867a1=(a=>{const b=d(a).body;return m(b)?W:l(b)});c.wbg.__wbg_appendChild_bd383ec5356c0bdb=function(){return C(((a,b)=>{const c=d(a).appendChild(d(b));return l(c)}),arguments)};c.wbg.__wbg_new_abda76e883ba8a5f=(()=>{const a=new V();return l(a)});c.wbg.__wbg_stack_658279fe44541cf6=((a,c)=>{const e=d(c).stack;const f=u(e,b.__wbindgen_malloc,b.__wbindgen_realloc);const g=r;q()[a/a5+ Y]=g;q()[a/a5+ W]=f});c.wbg.__wbg_error_f851667af71bcfc6=((a,c)=>{let d;let e;try{d=a;e=c;console.error(k(a,c))}finally{b.__wbindgen_free(d,e,Y)}});c.wbg.__wbg_resume_32fc64eaa464289a=function(){return C((a=>{const b=d(a).resume();return l(b)}),arguments)};c.wbg.__wbg_close_41fb3d66e34246dc=function(){return C((a=>{const b=d(a).close();return l(b)}),arguments)};c.wbg.__wbg_eval_6c93b88a4b3be0eb=function(){return C(((a,b)=>{const c=eval(k(a,b));return l(c)}),arguments)};c.wbg.__wbindgen_boolean_get=(a=>{const b=d(a);const c=typeof b===a0?(b?Y:W):2;return c});c.wbg.__wbg_crypto_d05b68a3572bb8ca=(a=>{const b=d(a).crypto;return l(b)});c.wbg.__wbindgen_is_object=(a=>{const b=d(a);const c=typeof b===`object`&&b!==T;return c});c.wbg.__wbg_process_b02b3570280d0366=(a=>{const b=d(a).process;return l(b)});c.wbg.__wbg_versions_c1cb42213cedf0f5=(a=>{const b=d(a).versions;return l(b)});c.wbg.__wbg_node_43b1089f407e4ec2=(a=>{const b=d(a).node;return l(b)});c.wbg.__wbindgen_is_string=(a=>{const b=typeof d(a)===a1;return b});c.wbg.__wbg_require_9a7e0f667ead4995=function(){return C((()=>{const a=module.require;return l(a)}),arguments)};c.wbg.__wbindgen_is_function=(a=>{const b=typeof d(a)===_;return b});c.wbg.__wbg_call_f6a2bc58c19c53c6=function(){return C(((a,b,c)=>{const e=d(a).call(d(b),d(c));return l(e)}),arguments)};c.wbg.__wbg_msCrypto_10fc94afee92bd76=(a=>{const b=d(a).msCrypto;return l(b)});c.wbg.__wbg_newwithlength_728575f3bba9959b=(a=>{const b=new X(a>>>W);return l(b)});c.wbg.__wbg_randomFillSync_b70ccbdf4926a99d=function(){return C(((a,b)=>{d(a).randomFillSync(g(b))}),arguments)};c.wbg.__wbg_subarray_7f7a652672800851=((a,b,c)=>{const e=d(a).subarray(b>>>W,c>>>W);return l(e)});c.wbg.__wbg_getRandomValues_7e42b4fb8779dc6d=function(){return C(((a,b)=>{d(a).getRandomValues(d(b))}),arguments)};c.wbg.__wbg_connected_e7984632fc5baeca=(a=>{const b=d(a).connected;return b});c.wbg.__wbg_now_86f7ca537c8b86d5=(()=>{const a=Date.now();return a});c.wbg.__wbindgen_number_get=((a,b)=>{const c=d(b);const e=typeof c===$?c:R;o()[a/8+ Y]=m(e)?W:e;q()[a/a5+ W]=!m(e)});c.wbg.__wbg_pressed_6689c2375fac761f=(a=>{const b=d(a).pressed;return b});c.wbg.__wbg_value_df73ea5624c95e17=(a=>{const b=d(a).value;return b});c.wbg.__wbindgen_is_null=(a=>{const b=d(a)===T;return b});c.wbg.__wbg_id_b65a9e5c38af3381=((a,c)=>{const e=d(c).id;const f=u(e,b.__wbindgen_malloc,b.__wbindgen_realloc);const g=r;q()[a/a5+ Y]=g;q()[a/a5+ W]=f});c.wbg.__wbg_buttons_42cf81a9dc81d560=(a=>{const b=d(a).buttons;return l(b)});c.wbg.__wbg_length_d99b680fd68bf71b=(a=>{const b=d(a).length;return b});c.wbg.__wbg_axes_29295fc5e2791881=(a=>{const b=d(a).axes;return l(b)});c.wbg.__wbg_mapping_9891f1fe2b08c14b=(a=>{const b=d(a).mapping;return l(b)});c.wbg.__wbg_get_c43534c00f382c8a=((a,b)=>{const c=d(a)[b>>>W];return l(c)});c.wbg.__wbg_isSecureContext_a4963ee69cd2dcf2=(a=>{const b=d(a).isSecureContext;return b});c.wbg.__wbg_navigator_910cca0226b70083=(a=>{const b=d(a).navigator;return l(b)});c.wbg.__wbg_getGamepads_77e0120579fb7c8d=function(){return C((a=>{const b=d(a).getGamepads();return l(b)}),arguments)};c.wbg.__wbg_index_ad8482877f40f04f=(a=>{const b=d(a).index;return b});c.wbg.__wbg_instanceof_DomException_3813d9e5e1312be8=(a=>{let b;try{b=d(a) instanceof DOMException}catch(a){b=!1}const c=b;return c});c.wbg.__wbg_message_9cb2b2d345ff18c6=((a,c)=>{const e=d(c).message;const f=u(e,b.__wbindgen_malloc,b.__wbindgen_realloc);const g=r;q()[a/a5+ Y]=g;q()[a/a5+ W]=f});c.wbg.__wbg_getExtension_25430e0ed157fcf8=function(){return C(((a,b,c)=>{const e=d(a).getExtension(k(b,c));return m(e)?W:l(e)}),arguments)};c.wbg.__wbg_getSupportedExtensions_1a007030d26efba5=(a=>{const b=d(a).getSupportedExtensions();return m(b)?W:l(b)});c.wbg.__wbg_getParameter_b282105ca8420119=function(){return C(((a,b)=>{const c=d(a).getParameter(b>>>W);return l(c)}),arguments)};c.wbg.__wbindgen_string_get=((a,c)=>{const e=d(c);const f=typeof e===a1?e:R;var g=m(f)?W:u(f,b.__wbindgen_malloc,b.__wbindgen_realloc);var h=r;q()[a/a5+ Y]=h;q()[a/a5+ W]=g});c.wbg.__wbg_texSubImage2D_4d372d780fc0e4a7=function(){return C(((a,b,c,e,f,g,h,i,j,k)=>{d(a).texSubImage2D(b>>>W,c,e,f,g,h,i>>>W,j>>>W,d(k))}),arguments)};c.wbg.__wbg_texSubImage2D_cbc346dc5a210f5d=function(){return C(((a,b,c,e,f,g,h,i,j,k)=>{d(a).texSubImage2D(b>>>W,c,e,f,g,h,i>>>W,j>>>W,d(k))}),arguments)};c.wbg.__wbg_texSubImage2D_ad0af504139d876c=function(){return C(((a,b,c,e,f,g,h,i,j,k)=>{d(a).texSubImage2D(b>>>W,c,e,f,g,h,i>>>W,j>>>W,d(k))}),arguments)};c.wbg.__wbg_texSubImage3D_e36d3c30ac0d0749=function(){return C(((a,b,c,e,f,g,h,i,j,k,l,m)=>{d(a).texSubImage3D(b>>>W,c,e,f,g,h,i,j,k>>>W,l>>>W,d(m))}),arguments)};c.wbg.__wbg_texSubImage3D_d98b6d6d4c3f3d01=function(){return C(((a,b,c,e,f,g,h,i,j,k,l,m)=>{d(a).texSubImage3D(b>>>W,c,e,f,g,h,i,j,k>>>W,l>>>W,d(m))}),arguments)};c.wbg.__wbg_texSubImage3D_a8f081c484f78039=function(){return C(((a,b,c,e,f,g,h,i,j,k,l,m)=>{d(a).texSubImage3D(b>>>W,c,e,f,g,h,i,j,k>>>W,l>>>W,d(m))}),arguments)};c.wbg.__wbg_framebufferTextureMultiviewOVR_b4f234dba08738d7=((a,b,c,e,f,g,h)=>{d(a).framebufferTextureMultiviewOVR(b>>>W,c>>>W,d(e),f,g,h)});c.wbg.__wbg_bindFramebuffer_28e8c0c3f76447af=((a,b,c)=>{d(a).bindFramebuffer(b>>>W,d(c))});c.wbg.__wbg_bindFramebuffer_0b8b88d70f0b876e=((a,b,c)=>{d(a).bindFramebuffer(b>>>W,d(c))});c.wbg.__wbg_getSupportedProfiles_a3af04122b4f2f30=(a=>{const b=d(a).getSupportedProfiles();return m(b)?W:l(b)});c.wbg.__wbg_new_34c624469fb1d4fd=(()=>{const a=new S();return l(a)});c.wbg.__wbg_includes_b0feae2b4a1ae514=((a,b,c)=>{const e=d(a).includes(d(b),c);return e});c.wbg.__wbg_createFramebuffer_f8c26154f8992bfa=(a=>{const b=d(a).createFramebuffer();return m(b)?W:l(b)});c.wbg.__wbg_createFramebuffer_52e5d7327d5afba3=(a=>{const b=d(a).createFramebuffer();return m(b)?W:l(b)});c.wbg.__wbg_createQuery_dca7163929abd29d=(a=>{const b=d(a).createQuery();return m(b)?W:l(b)});c.wbg.__wbg_createRenderbuffer_e2d77844fbdcc842=(a=>{const b=d(a).createRenderbuffer();return m(b)?W:l(b)});c.wbg.__wbg_createRenderbuffer_881be806709189a9=(a=>{const b=d(a).createRenderbuffer();return m(b)?W:l(b)});c.wbg.__wbg_createSampler_e2bcf2bc717a1cad=(a=>{const b=d(a).createSampler();return m(b)?W:l(b)});c.wbg.__wbg_createShader_896229165c5a11d4=((a,b)=>{const c=d(a).createShader(b>>>W);return m(c)?W:l(c)});c.wbg.__wbg_createShader_119ffcdb1667f405=((a,b)=>{const c=d(a).createShader(b>>>W);return m(c)?W:l(c)});c.wbg.__wbg_createTexture_b77eefdce0bb2c55=(a=>{const b=d(a).createTexture();return m(b)?W:l(b)});c.wbg.__wbg_createTexture_4f0c3c77df4bde11=(a=>{const b=d(a).createTexture();return m(b)?W:l(b)});c.wbg.__wbg_deleteShader_322b059ad560664a=((a,b)=>{d(a).deleteShader(d(b))});c.wbg.__wbg_deleteShader_5ec1e25476df2da0=((a,b)=>{d(a).deleteShader(d(b))});c.wbg.__wbg_shaderSource_04af20ecb1962b3b=((a,b,c,e)=>{d(a).shaderSource(d(b),k(c,e))});c.wbg.__wbg_shaderSource_e12efd3a2bf3413d=((a,b,c,e)=>{d(a).shaderSource(d(b),k(c,e))});c.wbg.__wbg_compileShader_be824cfad43331b8=((a,b)=>{d(a).compileShader(d(b))});c.wbg.__wbg_compileShader_bdfb3d5a3ad59498=((a,b)=>{d(a).compileShader(d(b))});c.wbg.__wbg_getShaderParameter_a9315ba73ab18731=((a,b,c)=>{const e=d(a).getShaderParameter(d(b),c>>>W);return l(e)});c.wbg.__wbg_getShaderParameter_58d3b34afa9db13b=((a,b,c)=>{const e=d(a).getShaderParameter(d(b),c>>>W);return l(e)});c.wbg.__wbg_createProgram_983b87cad6d06768=(a=>{const b=d(a).createProgram();return m(b)?W:l(b)});c.wbg.__wbg_createProgram_0a7670ed33f06d97=(a=>{const b=d(a).createProgram();return m(b)?W:l(b)});c.wbg.__wbg_deleteProgram_8447c337271aa934=((a,b)=>{d(a).deleteProgram(d(b))});c.wbg.__wbg_deleteProgram_d90e44574acb8018=((a,b)=>{d(a).deleteProgram(d(b))});c.wbg.__wbg_attachShader_cfbbdefc08a0422f=((a,b,c)=>{d(a).attachShader(d(b),d(c))});c.wbg.__wbg_attachShader_70c3f88b777a0a54=((a,b,c)=>{d(a).attachShader(d(b),d(c))});c.wbg.__wbg_linkProgram_caeab1eb0c0246be=((a,b)=>{d(a).linkProgram(d(b))});c.wbg.__wbg_linkProgram_e170ffe0b8242136=((a,b)=>{d(a).linkProgram(d(b))});c.wbg.__wbg_getProgramParameter_35800b92324ff726=((a,b,c)=>{const e=d(a).getProgramParameter(d(b),c>>>W);return l(e)});c.wbg.__wbg_getProgramParameter_22b3f1c8d913cd2c=((a,b,c)=>{const e=d(a).getProgramParameter(d(b),c>>>W);return l(e)});c.wbg.__wbg_getActiveUniform_07e2333f5c6a043b=((a,b,c)=>{const e=d(a).getActiveUniform(d(b),c>>>W);return m(e)?W:l(e)});c.wbg.__wbg_size_b9db0a8002af3240=(a=>{const b=d(a).size;return b});c.wbg.__wbg_type_d15cb90f6bbe57c6=(a=>{const b=d(a).type;return b});c.wbg.__wbg_name_4774375796143fca=((a,c)=>{const e=d(c).name;const f=u(e,b.__wbindgen_malloc,b.__wbindgen_realloc);const g=r;q()[a/a5+ Y]=g;q()[a/a5+ W]=f});c.wbg.__wbg_getActiveUniform_d5bafdccc7438fbb=((a,b,c)=>{const e=d(a).getActiveUniform(d(b),c>>>W);return m(e)?W:l(e)});c.wbg.__wbg_useProgram_53de6b084c4780ce=((a,b)=>{d(a).useProgram(d(b))});c.wbg.__wbg_useProgram_229c8fa8394b4c26=((a,b)=>{d(a).useProgram(d(b))});c.wbg.__wbg_createBuffer_90bf79c414ad4956=(a=>{const b=d(a).createBuffer();return m(b)?W:l(b)});c.wbg.__wbg_createBuffer_a95c59cc2c1750e7=(a=>{const b=d(a).createBuffer();return m(b)?W:l(b)});c.wbg.__wbg_bindBuffer_ac939bcab5249160=((a,b,c)=>{d(a).bindBuffer(b>>>W,d(c))});c.wbg.__wbg_bindBuffer_3f166cc2f502fc09=((a,b,c)=>{d(a).bindBuffer(b>>>W,d(c))});c.wbg.__wbg_bindBufferRange_bfdd227c2d5515af=((a,b,c,e,f,g)=>{d(a).bindBufferRange(b>>>W,c>>>W,d(e),f,g)});c.wbg.__wbg_bindRenderbuffer_f06f73fc0b43967e=((a,b,c)=>{d(a).bindRenderbuffer(b>>>W,d(c))});c.wbg.__wbg_bindRenderbuffer_2fe89083883b96e7=((a,b,c)=>{d(a).bindRenderbuffer(b>>>W,d(c))});c.wbg.__wbg_blitFramebuffer_8ca764978b2e3b3f=((a,b,c,e,f,g,h,i,j,k,l)=>{d(a).blitFramebuffer(b,c,e,f,g,h,i,j,k>>>W,l>>>W)});c.wbg.__wbg_createVertexArrayOES_96ccfea00081dcf3=(a=>{const b=d(a).createVertexArrayOES();return m(b)?W:l(b)});c.wbg.__wbg_createVertexArray_761ba22fc5da3ad7=(a=>{const b=d(a).createVertexArray();return m(b)?W:l(b)});c.wbg.__wbg_deleteVertexArray_26631f33de66bdfd=((a,b)=>{d(a).deleteVertexArray(d(b))});c.wbg.__wbg_deleteVertexArrayOES_657b2572282b9dff=((a,b)=>{d(a).deleteVertexArrayOES(d(b))});c.wbg.__wbg_bindVertexArray_2a70aed123353597=((a,b)=>{d(a).bindVertexArray(d(b))});c.wbg.__wbg_bindVertexArrayOES_e95cf32f50e47240=((a,b)=>{d(a).bindVertexArrayOES(d(b))});c.wbg.__wbg_pixelStorei_ac98844c2d6d1937=((a,b,c)=>{d(a).pixelStorei(b>>>W,c)});c.wbg.__wbg_pixelStorei_6be3fc7114b690b8=((a,b,c)=>{d(a).pixelStorei(b>>>W,c)});c.wbg.__wbg_bufferData_9566a2faddca5792=((a,b,c,e)=>{d(a).bufferData(b>>>W,c,e>>>W)});c.wbg.__wbg_bufferData_6c5edae24f952d4d=((a,b,c,e)=>{d(a).bufferData(b>>>W,c,e>>>W)});c.wbg.__wbg_bufferData_b2e68fdc1fd1e94b=((a,b,c,e)=>{d(a).bufferData(b>>>W,d(c),e>>>W)});c.wbg.__wbg_bufferData_eab63186e3e72d98=((a,b,c,e)=>{d(a).bufferData(b>>>W,d(c),e>>>W)});c.wbg.__wbg_bufferSubData_7d216abec8307331=((a,b,c,e)=>{d(a).bufferSubData(b>>>W,c,d(e))});c.wbg.__wbg_bufferSubData_3b75566851019327=((a,b,c,e)=>{d(a).bufferSubData(b>>>W,c,d(e))});c.wbg.__wbg_getBufferSubData_42fbdf01d4c31560=((a,b,c,e)=>{d(a).getBufferSubData(b>>>W,c,d(e))});c.wbg.__wbg_clearBufferiv_07046f3c028ef141=((a,b,c,e,f)=>{d(a).clearBufferiv(b>>>W,c,G(e,f))});c.wbg.__wbg_clearBufferuiv_d0ebea28b39eb980=((a,b,c,e,f)=>{d(a).clearBufferuiv(b>>>W,c,J(e,f))});c.wbg.__wbg_clearBufferfv_a415d907abca9e83=((a,b,c,e,f)=>{d(a).clearBufferfv(b>>>W,c,F(e,f))});c.wbg.__wbg_clearBufferfi_04f7c24deffc8e1d=((a,b,c,e,f)=>{d(a).clearBufferfi(b>>>W,c,e,f)});c.wbg.__wbg_clientWaitSync_b3f79a980d4d9498=((a,b,c,e)=>{const f=d(a).clientWaitSync(d(b),c>>>W,e>>>W);return f});c.wbg.__wbg_copyBufferSubData_70becf455ca484cd=((a,b,c,e,f,g)=>{d(a).copyBufferSubData(b>>>W,c>>>W,e,f,g)});c.wbg.__wbg_copyTexSubImage2D_6ce49c4a307e877d=((a,b,c,e,f,g,h,i,j)=>{d(a).copyTexSubImage2D(b>>>W,c,e,f,g,h,i,j)});c.wbg.__wbg_copyTexSubImage2D_6e2fe88bb9fa3ffd=((a,b,c,e,f,g,h,i,j)=>{d(a).copyTexSubImage2D(b>>>W,c,e,f,g,h,i,j)});c.wbg.__wbg_copyTexSubImage3D_f385cc4e05c95e64=((a,b,c,e,f,g,h,i,j,k)=>{d(a).copyTexSubImage3D(b>>>W,c,e,f,g,h,i,j,k)});c.wbg.__wbg_deleteBuffer_d70596808095dac2=((a,b)=>{d(a).deleteBuffer(d(b))});c.wbg.__wbg_deleteBuffer_b8aaa61f9bb13617=((a,b)=>{d(a).deleteBuffer(d(b))});c.wbg.__wbg_deleteFramebuffer_23c9c7c8176aa9b8=((a,b)=>{d(a).deleteFramebuffer(d(b))});c.wbg.__wbg_deleteFramebuffer_d6907809466bdbdb=((a,b)=>{d(a).deleteFramebuffer(d(b))});c.wbg.__wbg_deleteQuery_3524b489c741d48f=((a,b)=>{d(a).deleteQuery(d(b))});c.wbg.__wbg_deleteRenderbuffer_7bb3c4c79eac08ff=((a,b)=>{d(a).deleteRenderbuffer(d(b))});c.wbg.__wbg_deleteRenderbuffer_1c4b186beb91d4a5=((a,b)=>{d(a).deleteRenderbuffer(d(b))});c.wbg.__wbg_deleteSampler_f760c2bdc7a3d881=((a,b)=>{d(a).deleteSampler(d(b))});c.wbg.__wbg_deleteSync_6bff1584a3aae6a1=((a,b)=>{d(a).deleteSync(d(b))});c.wbg.__wbg_deleteTexture_bbda7cb554bc12b9=((a,b)=>{d(a).deleteTexture(d(b))});c.wbg.__wbg_deleteTexture_554c30847d340929=((a,b)=>{d(a).deleteTexture(d(b))});c.wbg.__wbg_disable_57e8624c865bd654=((a,b)=>{d(a).disable(b>>>W)});c.wbg.__wbg_disable_f68719f70ddfb5b8=((a,b)=>{d(a).disable(b>>>W)});c.wbg.__wbg_disableVertexAttribArray_fb822948cb54eec9=((a,b)=>{d(a).disableVertexAttribArray(b>>>W)});c.wbg.__wbg_disableVertexAttribArray_557393d91e187e24=((a,b)=>{d(a).disableVertexAttribArray(b>>>W)});c.wbg.__wbg_drawArrays_d48ee5c0a02be869=((a,b,c,e)=>{d(a).drawArrays(b>>>W,c,e)});c.wbg.__wbg_drawArrays_2f37c32534dffd91=((a,b,c,e)=>{d(a).drawArrays(b>>>W,c,e)});c.wbg.__wbg_drawArraysInstancedANGLE_4ba856b2c59d84b8=((a,b,c,e,f)=>{d(a).drawArraysInstancedANGLE(b>>>W,c,e,f)});c.wbg.__wbg_drawArraysInstanced_b0963fae97f2f14e=((a,b,c,e,f)=>{d(a).drawArraysInstanced(b>>>W,c,e,f)});c.wbg.__wbindgen_number_new=(a=>{const b=a;return l(b)});c.wbg.__wbg_push_906164999551d793=((a,b)=>{const c=d(a).push(d(b));return c});c.wbg.__wbg_of_3d7aa62bb0ab56ee=(a=>{const b=S.of(d(a));return l(b)});c.wbg.__wbg_drawBuffersWEBGL_533ee2b72ddb728a=((a,b)=>{d(a).drawBuffersWEBGL(d(b))});c.wbg.__wbg_drawBuffers_117fa4691357b53d=((a,b)=>{d(a).drawBuffers(d(b))});c.wbg.__wbg_drawElementsInstancedANGLE_fdf5cd2eb03dd141=((a,b,c,e,f,g)=>{d(a).drawElementsInstancedANGLE(b>>>W,c,e>>>W,f,g)});c.wbg.__wbg_drawElementsInstanced_19c02c2c6c7ebdd5=((a,b,c,e,f,g)=>{d(a).drawElementsInstanced(b>>>W,c,e>>>W,f,g)});c.wbg.__wbg_enable_54d01bacc240df3e=((a,b)=>{d(a).enable(b>>>W)});c.wbg.__wbg_enable_6dab9d5278ba15e2=((a,b)=>{d(a).enable(b>>>W)});c.wbg.__wbg_enableVertexAttribArray_c971ef03599058ec=((a,b)=>{d(a).enableVertexAttribArray(b>>>W)});c.wbg.__wbg_enableVertexAttribArray_c2bfb733e87824c8=((a,b)=>{d(a).enableVertexAttribArray(b>>>W)});c.wbg.__wbg_framebufferRenderbuffer_564b54a213de82b7=((a,b,c,e,f)=>{d(a).framebufferRenderbuffer(b>>>W,c>>>W,e>>>W,d(f))});c.wbg.__wbg_framebufferRenderbuffer_27bc520ea685b484=((a,b,c,e,f)=>{d(a).framebufferRenderbuffer(b>>>W,c>>>W,e>>>W,d(f))});c.wbg.__wbg_framebufferTexture2D_e7783c0015d1135a=((a,b,c,e,f,g)=>{d(a).framebufferTexture2D(b>>>W,c>>>W,e>>>W,d(f),g)});c.wbg.__wbg_framebufferTexture2D_c61bc6c888f33a52=((a,b,c,e,f,g)=>{d(a).framebufferTexture2D(b>>>W,c>>>W,e>>>W,d(f),g)});c.wbg.__wbg_framebufferTextureLayer_5fdc631245f13684=((a,b,c,e,f,g)=>{d(a).framebufferTextureLayer(b>>>W,c>>>W,d(e),f,g)});c.wbg.__wbg_frontFace_e13136966c974dd8=((a,b)=>{d(a).frontFace(b>>>W)});c.wbg.__wbg_frontFace_271693c85383f2e8=((a,b)=>{d(a).frontFace(b>>>W)});c.wbg.__wbg_getParameter_798cbb8ff20c7af0=function(){return C(((a,b)=>{const c=d(a).getParameter(b>>>W);return l(c)}),arguments)};c.wbg.__wbg_getIndexedParameter_69fe97ab84f9db9b=function(){return C(((a,b,c)=>{const e=d(a).getIndexedParameter(b>>>W,c>>>W);return l(e)}),arguments)};c.wbg.__wbg_getUniformLocation_f161344f25983444=((a,b,c,e)=>{const f=d(a).getUniformLocation(d(b),k(c,e));return m(f)?W:l(f)});c.wbg.__wbg_getUniformLocation_7b435a76db4f3128=((a,b,c,e)=>{const f=d(a).getUniformLocation(d(b),k(c,e));return m(f)?W:l(f)});c.wbg.__wbg_getSyncParameter_0c83093c52867612=((a,b,c)=>{const e=d(a).getSyncParameter(d(b),c>>>W);return l(e)});c.wbg.__wbg_renderbufferStorage_982fcb5577f764ca=((a,b,c,e,f)=>{d(a).renderbufferStorage(b>>>W,c>>>W,e,f)});c.wbg.__wbg_renderbufferStorage_5a63960c0bb41916=((a,b,c,e,f)=>{d(a).renderbufferStorage(b>>>W,c>>>W,e,f)});c.wbg.__wbg_renderbufferStorageMultisample_c5f884a4faf6330a=((a,b,c,e,f,g)=>{d(a).renderbufferStorageMultisample(b>>>W,c,e>>>W,f,g)});c.wbg.__wbg_samplerParameterf_a15f79d315dcfc5d=((a,b,c,e)=>{d(a).samplerParameterf(d(b),c>>>W,e)});c.wbg.__wbg_samplerParameteri_6f5c08b9c98433e9=((a,b,c,e)=>{d(a).samplerParameteri(d(b),c>>>W,e)});c.wbg.__wbg_texStorage2D_b46c4dcaa6dc9638=((a,b,c,e,f,g)=>{d(a).texStorage2D(b>>>W,c,e>>>W,f,g)});c.wbg.__wbg_texStorage3D_521eded8d8da33a6=((a,b,c,e,f,g,h)=>{d(a).texStorage3D(b>>>W,c,e>>>W,f,g,h)});c.wbg.__wbg_uniform1i_5a5f1f9d5828e6c6=((a,b,c)=>{d(a).uniform1i(d(b),c)});c.wbg.__wbg_uniform1i_1fd90743f7b78faa=((a,b,c)=>{d(a).uniform1i(d(b),c)});c.wbg.__wbg_uniform2iv_5ba0883cf01ae09d=((a,b,c,e)=>{d(a).uniform2iv(d(b),G(c,e))});c.wbg.__wbg_uniform2iv_fcef57681e7795f1=((a,b,c,e)=>{d(a).uniform2iv(d(b),G(c,e))});c.wbg.__wbg_uniform3iv_f297f19f134ad0c2=((a,b,c,e)=>{d(a).uniform3iv(d(b),G(c,e))});c.wbg.__wbg_uniform3iv_da537ca1568e83fe=((a,b,c,e)=>{d(a).uniform3iv(d(b),G(c,e))});c.wbg.__wbg_uniform4iv_2dbb8a34d36a28c3=((a,b,c,e)=>{d(a).uniform4iv(d(b),G(c,e))});c.wbg.__wbg_uniform4iv_eaebe8f50f18f893=((a,b,c,e)=>{d(a).uniform4iv(d(b),G(c,e))});c.wbg.__wbg_uniform1f_5c36f8a2cf1d8cd7=((a,b,c)=>{d(a).uniform1f(d(b),c)});c.wbg.__wbg_uniform1f_b13c736354a10aa7=((a,b,c)=>{d(a).uniform1f(d(b),c)});c.wbg.__wbg_uniform4f_93ef17b7172e8ad2=((a,b,c,e,f,g)=>{d(a).uniform4f(d(b),c,e,f,g)});c.wbg.__wbg_uniform4f_5b57101145ac6da8=((a,b,c,e,f,g)=>{d(a).uniform4f(d(b),c,e,f,g)});c.wbg.__wbg_uniform2fv_d375e6a7b2f1e575=((a,b,c,e)=>{d(a).uniform2fv(d(b),F(c,e))});c.wbg.__wbg_uniform2fv_a079de4d57adc89f=((a,b,c,e)=>{d(a).uniform2fv(d(b),F(c,e))});c.wbg.__wbg_uniform3fv_ce5f4b99b178dd74=((a,b,c,e)=>{d(a).uniform3fv(d(b),F(c,e))});c.wbg.__wbg_uniform3fv_0211c4807ed5b6bb=((a,b,c,e)=>{d(a).uniform3fv(d(b),F(c,e))});c.wbg.__wbg_uniform4fv_f7afb7d09ee03175=((a,b,c,e)=>{d(a).uniform4fv(d(b),F(c,e))});c.wbg.__wbg_uniform4fv_5134ae6d977cd056=((a,b,c,e)=>{d(a).uniform4fv(d(b),F(c,e))});c.wbg.__wbg_uniformMatrix2fv_9e0249ce783ce2be=((a,b,c,e,f)=>{d(a).uniformMatrix2fv(d(b),c!==W,F(e,f))});c.wbg.__wbg_uniformMatrix2fv_1c4f6d47a69eddf2=((a,b,c,e,f)=>{d(a).uniformMatrix2fv(d(b),c!==W,F(e,f))});c.wbg.__wbg_uniformMatrix3fv_2a9524cf34ecbd62=((a,b,c,e,f)=>{d(a).uniformMatrix3fv(d(b),c!==W,F(e,f))});c.wbg.__wbg_uniformMatrix3fv_5b337adcad4a038d=((a,b,c,e,f)=>{d(a).uniformMatrix3fv(d(b),c!==W,F(e,f))});c.wbg.__wbg_uniformMatrix4fv_4c466deaf158ed5b=((a,b,c,e,f)=>{d(a).uniformMatrix4fv(d(b),c!==W,F(e,f))});c.wbg.__wbg_uniformMatrix4fv_10075e61e88aea3b=((a,b,c,e,f)=>{d(a).uniformMatrix4fv(d(b),c!==W,F(e,f))});c.wbg.__wbg_cullFace_a65f5d17b1ff5686=((a,b)=>{d(a).cullFace(b>>>W)});c.wbg.__wbg_cullFace_68b06ff8967b93f3=((a,b)=>{d(a).cullFace(b>>>W)});c.wbg.__wbg_colorMask_57603facaeb6e2e3=((a,b,c,e,f)=>{d(a).colorMask(b!==W,c!==W,e!==W,f!==W)});c.wbg.__wbg_colorMask_743f2bbb6e3fb4e5=((a,b,c,e,f)=>{d(a).colorMask(b!==W,c!==W,e!==W,f!==W)});c.wbg.__wbg_depthMask_d2c08d83ea550563=((a,b)=>{d(a).depthMask(b!==W)});c.wbg.__wbg_depthMask_052a5e3afe45b590=((a,b)=>{d(a).depthMask(b!==W)});c.wbg.__wbg_blendColor_71d54d4dad7a369a=((a,b,c,e,f)=>{d(a).blendColor(b,c,e,f)});c.wbg.__wbg_blendColor_4416443539cdef95=((a,b,c,e,f)=>{d(a).blendColor(b,c,e,f)});c.wbg.__wbg_invalidateFramebuffer_2d3e8a1b99fd845c=function(){return C(((a,b,c)=>{d(a).invalidateFramebuffer(b>>>W,d(c))}),arguments)};c.wbg.__wbg_polygonOffset_442517f9de53e3de=((a,b,c)=>{d(a).polygonOffset(b,c)});c.wbg.__wbg_polygonOffset_0f2730043ba169b2=((a,b,c)=>{d(a).polygonOffset(b,c)});c.wbg.__wbg_bindTexture_e28115f3ea3da6d2=((a,b,c)=>{d(a).bindTexture(b>>>W,d(c))});c.wbg.__wbg_bindTexture_be92cdd3f162b4f9=((a,b,c)=>{d(a).bindTexture(b>>>W,d(c))});c.wbg.__wbg_bindSampler_12a1965a2db071ed=((a,b,c)=>{d(a).bindSampler(b>>>W,d(c))});c.wbg.__wbg_activeTexture_3748123e1becf07d=((a,b)=>{d(a).activeTexture(b>>>W)});c.wbg.__wbg_activeTexture_02d56293bce2f613=((a,b)=>{d(a).activeTexture(b>>>W)});c.wbg.__wbg_fenceSync_452ae6f3789bdc77=((a,b,c)=>{const e=d(a).fenceSync(b>>>W,c>>>W);return m(e)?W:l(e)});c.wbg.__wbg_texParameteri_dd08984388e62491=((a,b,c,e)=>{d(a).texParameteri(b>>>W,c>>>W,e)});c.wbg.__wbg_texParameteri_f5c0d085b77931dd=((a,b,c,e)=>{d(a).texParameteri(b>>>W,c>>>W,e)});c.wbg.__wbg_texSubImage2D_d23a3ec1fa60bdaf=function(){return C(((a,b,c,e,f,g,h,i,j,k)=>{d(a).texSubImage2D(b>>>W,c,e,f,g,h,i>>>W,j>>>W,d(k))}),arguments)};c.wbg.__wbg_texSubImage2D_33018bcf2de70890=function(){return C(((a,b,c,e,f,g,h,i,j,k)=>{d(a).texSubImage2D(b>>>W,c,e,f,g,h,i>>>W,j>>>W,d(k))}),arguments)};c.wbg.__wbg_texSubImage2D_b97aa5ddc0162314=function(){return C(((a,b,c,e,f,g,h,i,j,k)=>{d(a).texSubImage2D(b>>>W,c,e,f,g,h,i>>>W,j>>>W,k)}),arguments)};c.wbg.__wbg_compressedTexSubImage2D_5666e0146e152b7d=((a,b,c,e,f,g,h,i,j)=>{d(a).compressedTexSubImage2D(b>>>W,c,e,f,g,h,i>>>W,d(j))});c.wbg.__wbg_compressedTexSubImage2D_41270fc03b157293=((a,b,c,e,f,g,h,i,j)=>{d(a).compressedTexSubImage2D(b>>>W,c,e,f,g,h,i>>>W,d(j))});c.wbg.__wbg_compressedTexSubImage2D_1194f18adf8859b9=((a,b,c,e,f,g,h,i,j,k)=>{d(a).compressedTexSubImage2D(b>>>W,c,e,f,g,h,i>>>W,j,k)});c.wbg.__wbg_texSubImage3D_5884c8e282839ff9=function(){return C(((a,b,c,e,f,g,h,i,j,k,l,m)=>{d(a).texSubImage3D(b>>>W,c,e,f,g,h,i,j,k>>>W,l>>>W,m)}),arguments)};c.wbg.__wbg_texSubImage3D_2742ec6099cae3fc=function(){return C(((a,b,c,e,f,g,h,i,j,k,l,m)=>{d(a).texSubImage3D(b>>>W,c,e,f,g,h,i,j,k>>>W,l>>>W,d(m))}),arguments)};c.wbg.__wbg_compressedTexSubImage3D_34cd53cffc6add9c=((a,b,c,e,f,g,h,i,j,k,l,m)=>{d(a).compressedTexSubImage3D(b>>>W,c,e,f,g,h,i,j,k>>>W,l,m)});c.wbg.__wbg_compressedTexSubImage3D_f0da020d6e3e3791=((a,b,c,e,f,g,h,i,j,k,l)=>{d(a).compressedTexSubImage3D(b>>>W,c,e,f,g,h,i,j,k>>>W,d(l))});c.wbg.__wbg_depthFunc_b78eec6735fd7a0a=((a,b)=>{d(a).depthFunc(b>>>W)});c.wbg.__wbg_depthFunc_e49f522acf6c6c2d=((a,b)=>{d(a).depthFunc(b>>>W)});c.wbg.__wbg_depthRange_c4d7339e2f6b2e4a=((a,b,c)=>{d(a).depthRange(b,c)});c.wbg.__wbg_depthRange_8309e031492fd023=((a,b,c)=>{d(a).depthRange(b,c)});c.wbg.__wbg_scissor_7206bcd2a5540aa3=((a,b,c,e,f)=>{d(a).scissor(b,c,e,f)});c.wbg.__wbg_scissor_27cb154cc9864444=((a,b,c,e,f)=>{d(a).scissor(b,c,e,f)});c.wbg.__wbg_vertexAttribDivisorANGLE_51dd5c906f4912a2=((a,b,c)=>{d(a).vertexAttribDivisorANGLE(b>>>W,c>>>W)});c.wbg.__wbg_vertexAttribDivisor_aad38a21841ace46=((a,b,c)=>{d(a).vertexAttribDivisor(b>>>W,c>>>W)});c.wbg.__wbg_vertexAttribPointer_e9c4ff85658b9ad2=((a,b,c,e,f,g,h)=>{d(a).vertexAttribPointer(b>>>W,c,e>>>W,f!==W,g,h)});c.wbg.__wbg_vertexAttribPointer_3133080603a92d4c=((a,b,c,e,f,g,h)=>{d(a).vertexAttribPointer(b>>>W,c,e>>>W,f!==W,g,h)});c.wbg.__wbg_vertexAttribIPointer_24c9254053dd8ab4=((a,b,c,e,f,g)=>{d(a).vertexAttribIPointer(b>>>W,c,e>>>W,f,g)});c.wbg.__wbg_viewport_0ca27d1d6ac8424c=((a,b,c,e,f)=>{d(a).viewport(b,c,e,f)});c.wbg.__wbg_viewport_afd5166081d009b2=((a,b,c,e,f)=>{d(a).viewport(b,c,e,f)});c.wbg.__wbg_blendFunc_5adf0a3a9f164e6e=((a,b,c)=>{d(a).blendFunc(b>>>W,c>>>W)});c.wbg.__wbg_blendFunc_ac53b0d3a97b7f7f=((a,b,c)=>{d(a).blendFunc(b>>>W,c>>>W)});c.wbg.__wbg_blendFuncSeparate_52fdb0f1fbf57928=((a,b,c,e,f)=>{d(a).blendFuncSeparate(b>>>W,c>>>W,e>>>W,f>>>W)});c.wbg.__wbg_blendFuncSeparate_b6a96b8e26e75171=((a,b,c,e,f)=>{d(a).blendFuncSeparate(b>>>W,c>>>W,e>>>W,f>>>W)});c.wbg.__wbg_blendEquation_b1df5434f3ad5aac=((a,b)=>{d(a).blendEquation(b>>>W)});c.wbg.__wbg_blendEquation_1c7272d8e9e0ce11=((a,b)=>{d(a).blendEquation(b>>>W)});c.wbg.__wbg_blendEquationSeparate_33f23a57d77e8079=((a,b,c)=>{d(a).blendEquationSeparate(b>>>W,c>>>W)});c.wbg.__wbg_blendEquationSeparate_457e81472270e23c=((a,b,c)=>{d(a).blendEquationSeparate(b>>>W,c>>>W)});c.wbg.__wbg_stencilFuncSeparate_89563ca030dab790=((a,b,c,e,f)=>{d(a).stencilFuncSeparate(b>>>W,c>>>W,e,f>>>W)});c.wbg.__wbg_stencilFuncSeparate_e6b4c784aa470ba1=((a,b,c,e,f)=>{d(a).stencilFuncSeparate(b>>>W,c>>>W,e,f>>>W)});c.wbg.__wbg_stencilMask_76ea69a0c4738423=((a,b)=>{d(a).stencilMask(b>>>W)});c.wbg.__wbg_stencilMask_4093c371489c5e3e=((a,b)=>{d(a).stencilMask(b>>>W)});c.wbg.__wbg_stencilMaskSeparate_1303b1855315b85a=((a,b,c)=>{d(a).stencilMaskSeparate(b>>>W,c>>>W)});c.wbg.__wbg_stencilMaskSeparate_6a90a6801f96c33e=((a,b,c)=>{d(a).stencilMaskSeparate(b>>>W,c>>>W)});c.wbg.__wbg_stencilOpSeparate_fef362ec0f1539d1=((a,b,c,e,f)=>{d(a).stencilOpSeparate(b>>>W,c>>>W,e>>>W,f>>>W)});c.wbg.__wbg_stencilOpSeparate_f98bb31212170061=((a,b,c,e,f)=>{d(a).stencilOpSeparate(b>>>W,c>>>W,e>>>W,f>>>W)});c.wbg.__wbg_getUniformBlockIndex_b9628e75250e866c=((a,b,c,e)=>{const f=d(a).getUniformBlockIndex(d(b),k(c,e));return f});c.wbg.__wbg_uniformBlockBinding_bcbb7fbc2fe88b8d=((a,b,c,e)=>{d(a).uniformBlockBinding(d(b),c>>>W,e>>>W)});c.wbg.__wbg_readBuffer_4c16fe804e5fd30c=((a,b)=>{d(a).readBuffer(b>>>W)});c.wbg.__wbg_readPixels_8260b74d4439418e=function(){return C(((a,b,c,e,f,g,h,i)=>{d(a).readPixels(b,c,e,f,g>>>W,h>>>W,i)}),arguments)};c.wbg.__wbg_readPixels_c1a5f8a1344005bd=function(){return C(((a,b,c,e,f,g,h,i)=>{d(a).readPixels(b,c,e,f,g>>>W,h>>>W,d(i))}),arguments)};c.wbg.__wbg_readPixels_32bab95664f5bcdf=function(){return C(((a,b,c,e,f,g,h,i)=>{d(a).readPixels(b,c,e,f,g>>>W,h>>>W,d(i))}),arguments)};c.wbg.__wbg_beginQuery_ad59d7ffda61cf9e=((a,b,c)=>{d(a).beginQuery(b>>>W,d(c))});c.wbg.__wbg_endQuery_feb28d278e32cfae=((a,b)=>{d(a).endQuery(b>>>W)});c.wbg.__wbg_getQueryParameter_112c9a3c8a8dd0da=((a,b,c)=>{const e=d(a).getQueryParameter(d(b),c>>>W);return l(e)});c.wbg.__wbg_get_5027b32da70f39b1=function(){return C(((a,b)=>{const c=a8.get(d(a),d(b));return l(c)}),arguments)};c.wbg.__wbg_now_65ff8ec2b863300c=(a=>{const b=d(a).now();return b});c.wbg.__wbg_self_086b5302bcafb962=function(){return C((()=>{const a=self.self;return l(a)}),arguments)};c.wbg.__wbg_window_132fa5d7546f1de5=function(){return C((()=>{const a=window.window;return l(a)}),arguments)};c.wbg.__wbg_globalThis_e5f801a37ad7d07b=function(){return C((()=>{const a=globalThis.globalThis;return l(a)}),arguments)};c.wbg.__wbg_global_f9a61fce4af6b7c1=function(){return C((()=>{const a=global.global;return l(a)}),arguments)};c.wbg.__wbindgen_is_undefined=(a=>{const b=d(a)===R;return b});c.wbg.__wbg_newnoargs_5859b6d41c6fe9f7=((a,b)=>{const c=new Function(k(a,b));return l(c)});c.wbg.__wbg_call_a79f1973a4f07d5e=function(){return C(((a,b)=>{const c=d(a).call(d(b));return l(c)}),arguments)};c.wbg.__wbg_resolve_97ecd55ee839391b=(a=>{const b=Promise.resolve(d(a));return l(b)});c.wbg.__wbg_then_7aeb7c5f1536640f=((a,b)=>{const c=d(a).then(d(b));return l(c)});c.wbg.__wbg_then_5842e4e97f7beace=((a,b,c)=>{const e=d(a).then(d(b),d(c));return l(e)});c.wbg.__wbindgen_memory=(()=>{const a=b.memory;return l(a)});c.wbg.__wbg_buffer_5d1b598a01b41a42=(a=>{const b=d(a).buffer;return l(b)});c.wbg.__wbg_newwithbyteoffsetandlength_54c7b98977affdec=((a,b,c)=>{const e=new Int8Array(d(a),b>>>W,c>>>W);return l(e)});c.wbg.__wbg_newwithbyteoffsetandlength_16ba6d10861ea013=((a,b,c)=>{const e=new Int16Array(d(a),b>>>W,c>>>W);return l(e)});c.wbg.__wbg_newwithbyteoffsetandlength_821c7736f0d22b04=((a,b,c)=>{const e=new Z(d(a),b>>>W,c>>>W);return l(e)});c.wbg.__wbg_newwithbyteoffsetandlength_d695c7957788f922=((a,b,c)=>{const e=new X(d(a),b>>>W,c>>>W);return l(e)});c.wbg.__wbg_set_74906aa30864df5a=((a,b,c)=>{d(a).set(d(b),c>>>W)});c.wbg.__wbg_length_f0764416ba5bb237=(a=>{const b=d(a).length;return b});c.wbg.__wbg_newwithbyteoffsetandlength_2412e38a0385bbe2=((a,b,c)=>{const e=new Uint16Array(d(a),b>>>W,c>>>W);return l(e)});c.wbg.__wbg_newwithbyteoffsetandlength_aeed38cac7555df7=((a,b,c)=>{const e=new a6(d(a),b>>>W,c>>>W);return l(e)});c.wbg.__wbg_newwithbyteoffsetandlength_21163b4dfcbc673c=((a,b,c)=>{const e=new a4(d(a),b>>>W,c>>>W);return l(e)});c.wbg.__wbg_set_37a50e901587b477=function(){return C(((a,b,c)=>{const e=a8.set(d(a),d(b),d(c));return e}),arguments)};c.wbg.__wbg_newwithcontextoptions_14f6b0728f2e5974=function(){return C((b=>{const c=new a(d(b));return l(c)}),arguments)};c.wbg.__wbg_destination_62d2e29a54544ec0=(a=>{const b=d(a).destination;return l(b)});c.wbg.__wbg_maxChannelCount_f7897fa7dc85d572=(a=>{const b=d(a).maxChannelCount;return b});c.wbg.__wbg_setchannelCount_ddf571b2ad2e8eef=((a,b)=>{d(a).channelCount=b>>>W});c.wbg.__wbg_createBuffer_a4cdfb0b3c256e3e=function(){return C(((a,b,c,e)=>{const f=d(a).createBuffer(b>>>W,c>>>W,e);return l(f)}),arguments)};c.wbg.__wbg_currentTime_6b9141913a965d2f=(a=>{const b=d(a).currentTime;return b});c.wbg.__wbg_createBufferSource_0d20dc119e4ded11=function(){return C((a=>{const b=d(a).createBufferSource();return l(b)}),arguments)};c.wbg.__wbg_setbuffer_f16a95796c5a7380=((a,b)=>{d(a).buffer=d(b)});c.wbg.__wbg_connect_65474f2479b77506=function(){return C(((a,b)=>{const c=d(a).connect(d(b));return l(c)}),arguments)};c.wbg.__wbg_setonended_d2cab878358a6af4=((a,b)=>{d(a).onended=d(b)});c.wbg.__wbg_start_88dbb78b1c762033=function(){return C(((a,b)=>{d(a).start(b)}),arguments)};c.wbg.__wbg_copyToChannel_47042ca9c7b9618d=function(){return C(((a,b,c,e)=>{d(a).copyToChannel(F(b,c),e)}),arguments)};c.wbg.__wbg_measure_aa7a73f17813f708=function(){return C(((a,c,d,e)=>{let f;let g;let h;let i;try{f=a;g=c;h=d;i=e;performance.measure(k(a,c),k(d,e))}finally{b.__wbindgen_free(f,g,Y);b.__wbindgen_free(h,i,Y)}}),arguments)};c.wbg.__wbindgen_debug_string=((a,c)=>{const e=v(d(c));const f=u(e,b.__wbindgen_malloc,b.__wbindgen_realloc);const g=r;q()[a/a5+ Y]=g;q()[a/a5+ W]=f});c.wbg.__wbindgen_throw=((a,b)=>{throw new V(k(a,b))});c.wbg.__wbg_queueMicrotask_118eeb525d584d9a=(a=>{queueMicrotask(d(a))});c.wbg.__wbg_queueMicrotask_26a89c14c53809c0=(a=>{const b=d(a).queueMicrotask;return l(b)});c.wbg.__wbg_instanceof_WebGl2RenderingContext_92adf5bbd2568b71=(a=>{let b;try{b=d(a) instanceof WebGL2RenderingContext}catch(a){b=!1}const c=b;return c});c.wbg.__wbg_getProgramInfoLog_110f43b4125782e9=((a,c,e)=>{const f=d(c).getProgramInfoLog(d(e));var g=m(f)?W:u(f,b.__wbindgen_malloc,b.__wbindgen_realloc);var h=r;q()[a/a5+ Y]=h;q()[a/a5+ W]=g});c.wbg.__wbg_getShaderInfoLog_562b1447e7c24866=((a,c,e)=>{const f=d(c).getShaderInfoLog(d(e));var g=m(f)?W:u(f,b.__wbindgen_malloc,b.__wbindgen_realloc);var h=r;q()[a/a5+ Y]=h;q()[a/a5+ W]=g});c.wbg.__wbg_instanceof_Window_99dc9805eaa2614b=(a=>{let b;try{b=d(a) instanceof Window}catch(a){b=!1}const c=b;return c});c.wbg.__wbg_innerWidth_0bae627f0302b204=function(){return C((a=>{const b=d(a).innerWidth;return l(b)}),arguments)};c.wbg.__wbg_innerHeight_dc4c81e04e8bc294=function(){return C((a=>{const b=d(a).innerHeight;return l(b)}),arguments)};c.wbg.__wbg_devicePixelRatio_93bac98af723c7ba=(a=>{const b=d(a).devicePixelRatio;return b});c.wbg.__wbg_open_0aa18467f0bb625e=function(){return C(((a,b,c,e,f)=>{const g=d(a).open(k(b,c),k(e,f));return m(g)?W:l(g)}),arguments)};c.wbg.__wbg_get_7f8cd08c0360178c=((a,b,c)=>{const e=d(a)[k(b,c)];return m(e)?W:l(e)});c.wbg.__wbg_cancelAnimationFrame_2635bb6bdb94eb3f=function(){return C(((a,b)=>{d(a).cancelAnimationFrame(b)}),arguments)};c.wbg.__wbg_clearTimeout_cf250b4eed087f7b=((a,b)=>{d(a).clearTimeout(b)});c.wbg.__wbg_fullscreenElement_8ebe202aecd8ae3c=(a=>{const b=d(a).fullscreenElement;return m(b)?W:l(b)});c.wbg.__wbg_createElement_1a136faad4101f43=function(){return C(((a,b,c)=>{const e=d(a).createElement(k(b,c));return l(e)}),arguments)};c.wbg.__wbg_exitFullscreen_4207821551a694a7=(a=>{d(a).exitFullscreen()});c.wbg.__wbg_exitPointerLock_5764ae8300935b9e=(a=>{d(a).exitPointerLock()});c.wbg.__wbg_requestPointerLock_7e461e6d3998ca6a=(a=>{d(a).requestPointerLock()});c.wbg.__wbg_setAttribute_0918ea45d5a1c663=function(){return C(((a,b,c,e,f)=>{d(a).setAttribute(k(b,c),k(e,f))}),arguments)};c.wbg.__wbg_getProgramInfoLog_3ff10ea818ab6ce4=((a,c,e)=>{const f=d(c).getProgramInfoLog(d(e));var g=m(f)?W:u(f,b.__wbindgen_malloc,b.__wbindgen_realloc);var h=r;q()[a/a5+ Y]=h;q()[a/a5+ W]=g});c.wbg.__wbg_getShaderInfoLog_3e435d2b50e0ecf0=((a,c,e)=>{const f=d(c).getShaderInfoLog(d(e));var g=m(f)?W:u(f,b.__wbindgen_malloc,b.__wbindgen_realloc);var h=r;q()[a/a5+ Y]=h;q()[a/a5+ W]=g});c.wbg.__wbg_style_b32d5cb9a6bd4720=(a=>{const b=d(a).style;return l(b)});c.wbg.__wbg_ctrlKey_0d75e0e9028bd999=(a=>{const b=d(a).ctrlKey;return b});c.wbg.__wbg_shiftKey_12353f0e19b21d6a=(a=>{const b=d(a).shiftKey;return b});c.wbg.__wbg_altKey_a076f8612103d7e8=(a=>{const b=d(a).altKey;return b});c.wbg.__wbg_metaKey_4e3f6e986f2802b1=(a=>{const b=d(a).metaKey;return b});c.wbg.__wbg_button_8a97c55db17c7314=(a=>{const b=d(a).button;return b});c.wbg.__wbg_setProperty_a763529f4ef8ac76=function(){return C(((a,b,c,e,f)=>{d(a).setProperty(k(b,c),k(e,f))}),arguments)};c.wbg.__wbg_width_05e7fce75535d85f=(a=>{const b=d(a).width;return b});c.wbg.__wbg_height_51b9308e888df865=(a=>{const b=d(a).height;return b});c.wbg.__wbg_width_193b434156effb1d=(a=>{const b=d(a).width;return b});c.wbg.__wbg_setwidth_62ca8c8f2794be77=((a,b)=>{d(a).width=b>>>W});c.wbg.__wbg_height_84d4ae4d422188a3=(a=>{const b=d(a).height;return b});c.wbg.__wbg_setheight_34b71cfdf6095cbd=((a,b)=>{d(a).height=b>>>W});c.wbg.__wbg_addListener_0bbd0358c52d8a0e=function(){return C(((a,b)=>{d(a).addListener(d(b))}),arguments)};c.wbg.__wbg_removeListener_b8fc928c2300e3c6=function(){return C(((a,b)=>{d(a).removeListener(d(b))}),arguments)};c.wbg.__wbg_x_a0dc9ae3ec018f6a=(a=>{const b=d(a).x;return b});c.wbg.__wbg_y_6f1256869a5628ee=(a=>{const b=d(a).y;return b});c.wbg.__wbg_shiftKey_0a061aeba25dbd63=(a=>{const b=d(a).shiftKey;return b});c.wbg.__wbg_metaKey_b879a69fa9f3f7af=(a=>{const b=d(a).metaKey;return b});c.wbg.__wbg_code_3b51bddc7419ef7d=((a,c)=>{const e=d(c).code;const f=u(e,b.__wbindgen_malloc,b.__wbindgen_realloc);const g=r;q()[a/a5+ Y]=g;q()[a/a5+ W]=f});c.wbg.__wbg_deltaX_de18e6f358ab88cf=(a=>{const b=d(a).deltaX;return b});c.wbg.__wbg_deltaY_50a026b7421f883d=(a=>{const b=d(a).deltaY;return b});c.wbg.__wbg_deltaMode_b8290e36698673d0=(a=>{const b=d(a).deltaMode;return b});c.wbg.__wbg_setwidth_05075fb6b4cc720e=((a,b)=>{d(a).width=b>>>W});c.wbg.__wbg_setheight_7e0e88a922100d8c=((a,b)=>{d(a).height=b>>>W});c.wbg.__wbg_getContext_1daf9aba3e114993=function(){return C(((a,b,c,e)=>{const f=d(a).getContext(k(b,c),d(e));return m(f)?W:l(f)}),arguments)};c.wbg.__wbg_videoWidth_024256de61021e4a=(a=>{const b=d(a).videoWidth;return b});c.wbg.__wbg_videoHeight_2c601663d2d0211a=(a=>{const b=d(a).videoHeight;return b});c.wbg.__wbindgen_closure_wrapper3192=((a,b,c)=>{const d=w(a,b,2249,x);return l(d)});c.wbg.__wbindgen_closure_wrapper60448=((a,b,c)=>{const d=w(a,b,a9,y);return l(d)});c.wbg.__wbindgen_closure_wrapper60450=((a,b,c)=>{const d=w(a,b,a9,y);return l(d)});c.wbg.__wbindgen_closure_wrapper60452=((a,b,c)=>{const d=w(a,b,a9,y);return l(d)});c.wbg.__wbindgen_closure_wrapper60454=((a,b,c)=>{const d=w(a,b,a9,y);return l(d)});c.wbg.__wbindgen_closure_wrapper60456=((a,b,c)=>{const d=w(a,b,a9,y);return l(d)});c.wbg.__wbindgen_closure_wrapper60458=((a,b,c)=>{const d=w(a,b,a9,y);return l(d)});c.wbg.__wbindgen_closure_wrapper60460=((a,b,c)=>{const d=w(a,b,a9,y);return l(d)});c.wbg.__wbindgen_closure_wrapper69820=((a,b,c)=>{const d=w(a,b,44968,z);return l(d)});c.wbg.__wbindgen_closure_wrapper71534=((a,b,c)=>{const d=w(a,b,45347,A);return l(d)});c.wbg.__wbindgen_closure_wrapper75644=((a,b,c)=>{const d=w(a,b,46344,B);return l(d)});return c});var I=(()=>{if(H===T||H.byteLength===W){H=new a6(b.memory.buffer)};return H});var u=((a,b,c)=>{if(c===R){const c=s.encode(a);const d=b(c.length,Y)>>>W;j().subarray(d,d+ c.length).set(c);r=c.length;return d};let d=a.length;let e=b(d,Y)>>>W;const f=j();let g=W;for(;g<d;g++){const b=a.charCodeAt(g);if(b>127)break;f[e+ g]=b};if(g!==d){if(g!==W){a=a.slice(g)};e=c(e,d,d=g+ a.length*3,Y)>>>W;const b=j().subarray(e+ g,e+ d);const f=t(a,b);g+=f.written};r=g;return e});var j=(()=>{if(i===T||i.byteLength===W){i=new X(b.memory.buffer)};return i});var K=(async(a,b)=>{if(typeof Response===_&&a instanceof Response){if(typeof WebAssembly.instantiateStreaming===_){try{return await WebAssembly.instantiateStreaming(a,b)}catch(b){if(a.headers.get(`Content-Type`)!=`application/wasm`){console.warn(`\`WebAssembly.instantiateStreaming\` failed because your server does not serve wasm with \`application/wasm\` MIME type. Falling back to \`WebAssembly.instantiate\` which is slower. Original error:\\n`,b)}else{throw b}}};const c=await a.arrayBuffer();return await WebAssembly.instantiate(c,b)}else{const c=await WebAssembly.instantiate(a,b);if(c instanceof WebAssembly.Instance){return {instance:c,module:a}}else{return c}}});var G=((a,b)=>{a=a>>>W;return q().subarray(a/a5,a/a5+ b)});var J=((a,b)=>{a=a>>>W;return I().subarray(a/a5,a/a5+ b)});var z=((a,c)=>{b.wasm_bindgen__convert__closures__invoke0_mut__hc09e9020f5ee71bd(a,c)});var l=(a=>{if(e===c.length)c.push(c.length+ Y);const b=e;e=c[b];c[b]=a;return b});var g=(a=>{const b=d(a);f(a);return b});var q=(()=>{if(p===T||p.byteLength===W){p=new Z(b.memory.buffer)};return p});var B=((a,c,d)=>{b._dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__h67c6c39025c42490(a,c,l(d))});var P=(async(a)=>{if(b!==R)return b;if(typeof a===Q){a=new URL(`egui-gizmo-demo_bg.wasm`,import.meta.url)};const c=L();if(typeof a===a1||typeof Request===_&&a instanceof Request||typeof URL===_&&a instanceof URL){a=fetch(a)};M(c);const {instance:d,module:e}=await K(await a,c);return N(d,e)});var A=((a,c,d)=>{b._dyn_core__ops__function__FnMut__A____Output___R_as_wasm_bindgen__closure__WasmClosure___describe__invoke__had7d730985a1fbbd(a,c,l(d))});var f=(a=>{if(a<132)return;c[a]=e;e=a});var O=(a=>{if(b!==R)return b;const c=L();M(c);if(!(a instanceof WebAssembly.Module)){a=new WebAssembly.Module(a)};const d=new WebAssembly.Instance(a,c);return N(d,a)});var N=((a,c)=>{b=a.exports;P.__wbindgen_wasm_module=c;D=T;n=T;p=T;H=T;i=T;b.__wbindgen_start();return b});var w=((a,c,d,e)=>{const f={a:a,b:c,cnt:Y,dtor:d};const g=(...a)=>{f.cnt++;const c=f.a;f.a=W;try{return e(c,f.b,...a)}finally{if(--f.cnt===W){b.__wbindgen_export_2.get(f.dtor)(c,f.b)}else{f.a=c}}};g.original=f;return g});var k=((a,b)=>{a=a>>>W;return h.decode(j().subarray(a,a+ b))});var x=((a,c)=>{b.wasm_bindgen__convert__closures__invoke0_mut__h56db7e0444ea06f0(a,c)});const a=typeof AudioContext!==Q?AudioContext:(typeof webkitAudioContext!==Q?webkitAudioContext:R);let b;const c=new S(128).fill(R);c.push(R,T,!0,!1);let e=c.length;const h=typeof TextDecoder!==Q?new TextDecoder(U,{ignoreBOM:!0,fatal:!0}):{decode:()=>{throw V(`TextDecoder not available`)}};if(typeof TextDecoder!==Q){h.decode()};let i=T;let n=T;let p=T;let r=W;const s=typeof TextEncoder!==Q?new TextEncoder(U):{encode:()=>{throw V(`TextEncoder not available`)}};const t=typeof s.encodeInto===_?((a,b)=>s.encodeInto(a,b)):((a,b)=>{const c=s.encode(a);b.set(c);return {read:a.length,written:c.length}});let D=T;let H=T;export default P;export{O as initSync}
