--- docs/index.html
@@ -1,1 +1,0 @@
-<title>egui-gizmo demo</title><style>body,canvas,html{margin:0;padding:0;width:100%;height:100%;overflow:hidden;position:absolute;background:black;z-index:0}</style><link as=fetch crossorigin href=./egui-gizmo-demo_bg.wasm integrity=sha384-CoZgxOfGNgMidVxJZBj4y6IuxI1sBygjdJhVTE1STBBbuI7MwBz-Db2lvBAIZn27 rel=preload type=application/wasm><link crossorigin href=./egui-gizmo-demo.js integrity=sha384-pWDv33DuDrrkUHe2XL4zSWeCkMKnGRdKnj4ZFZtrJ9rHUbhYIC37AQfABk_zoZ-C rel=modulepreload></head><body oncontextmenu="return false;"><script type=module>import a,*as b from"./egui-gizmo-demo.js";a(`./egui-gizmo-demo_bg.wasm`);window.wasmBindings=b</script></body></html>
